Add minishift detection to demo
As of minishift 1.7.0 it is also able to support KubeVirt. This change
adds detection to determine whether minishift is installed/running as an
alternative to minikube, and uses whichever is "up".

This currently does however make the assumption that if either is up then it is
the desired target for kubevirt AND what kubectl is configured to use. If both
are up, which is possible on the same system, then it will use whatever
installation kubectl is currently configured for.